fix: [#601] bug in tags and categories filters
Going back with the browser button in the torrent list page produces an
URL with a param like this "tagFilters=os" when the previous filters only contain sone item.

That can be interpreted as a flat option or an array with one element.

The application was applying the first option and the index-api-lib
package expects the tagFilters to be always an array of strings.

Same for categoryFilters.

This only happended using the "back" buttom.